#56 Migration for materials management in NataliaLM/TFMGoSkiii#56 Total of Material Reservation needs to be decimal.

+++ TFMGoSki/Models/MaterialReservation.cs
@@ -6,15 +6,15 @@ namespace TFMGoSki.Models
     {
         public int Id { get; set; }
         public int UserId { get; set; }
-        public int Total { get; set; }
+        public decimal Total { get; set; }
         public bool Paid { get; set; }
-        public MaterialReservation(int userId, int total, bool paid)
+        public MaterialReservation(int userId, decimal total, bool paid)
         {
             UserId = userId;
             Total = total;
             Paid = paid;
         }
-        public MaterialReservation Update(int userId, int total, bool paid)
+        public MaterialReservation Update(int userId, decimal total, bool paid)
         {
             UserId = userId;
             Total = total;
